typ klar med allt
